update: modify Brewfile, enhance link_config_for_mac.sh, and update tmux and wezterm configurations

+++ link_config_for_mac.sh
@@ -1,8 +1,37 @@
+# Install brew
+/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+eval "$(/opt/homebrew/bin/brew shellenv)"
+brew bundle --file=./brew/Brewfile
+
+# Oh my zsh
+sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+
+# tmux theme
+LIB_FILE="$(pwd)/tmux/lib.txt"
+echo "🚀 Checking lib theme for tmux"
+if [ ! -f "$LIB_FILE" ]; then
+    echo "❌ Error: not found $LIB_FILE!"
+    exit 1
+fi
+
+while IFS= read -r line || [ -n "$line" ]; do
+    # Bỏ qua các dòng trống hoặc dòng bắt đầu bằng dấu #
+    [[ -z "$line" || "$line" =~ ^# ]] && continue
+    
+    # Thực thi lệnh git clone từ file lib.txt 
+    echo "> Chạy: $line"
+    eval "$line"
+done < "$LIB_FILE"
+
+# Link file 
+echo "🚀 Linking configuration files"
 ln -sf "$(pwd)/zsh/config/.zshrc" ~/.zshrc
 ln -sf "$(pwd)/zsh/powerlevel10k/.p10k.zsh" ~/.p10k.zsh
-ln -sf "$(pwd)/wezterm/.wezterm.lua" ~/.wezterm.lua
+ln -sf "$(pwd)/wezterm/wezterm.lua" ~/.wezterm.lua
 ln -sf "$(pwd)/tmux/tmux.conf" ~/.tmux.conf
 ln -sf "$(pwd)/.opencode" ~/.opencode
+
+source ~/.zshrc
 
 chmod +x ./ssh/server_access_ssh.sh           # đảm bảo có quyền thực thi
 ln -s "$(pwd)/ssh/server_access_ssh.sh" ~/server_access_ssh.sh
